+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.6.0 – Upcoming
+ - Remove `namespace` option for File
  - Add builder pattern to condense configuration
 
    ```